update impacted tests and sample code
